feature: restructure
