+++ app/src/main/java/game/com/twentyfour/containers/play/GameContract.kt
@@ -1,0 +1,7 @@
+package game.com.twentyfour.containers.play
+
+/**
+ * Created by yanyunhuang on 17/04/2018.
+ */
+interface GameContract {
+}
