+++ app/src/main/java/game/com/twentyfour/containers/play/GameFragment.kt
@@ -1,0 +1,8 @@
+package game.com.twentyfour.containers.play
+
+/**
+ * all views should be shown here
+ * Created by yanyunhuang on 17/04/2018.
+ */
+class GameFragment {
+}
