+++ app/src/main/java/game/com/twentyfour/containers/play/GamePrsenter.kt
@@ -1,0 +1,7 @@
+package game.com.twentyfour.containers.play
+
+/**
+ * Created by yanyunhuang on 17/04/2018.
+ */
+class GamePrsenter {
+}
