+++ app/src/main/java/game/com/twentyfour/models/Suits.kt
@@ -1,0 +1,8 @@
+package game.com.twentyfour.models
+
+/**
+ * Poker suits
+ * Created by yanyunhuang on 17/04/2018.
+ */
+enum class Suits {
+}
