+++ app/src/main/res/layout/activity_start.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#0099cc"
-    tools:context="game.com.twenty_four.StartActivity">
+    tools:context="game.com.twenty_four.game.com.twentyfour.containers.StartActivity">
 
     <TextView
         android:id="@+id/ready"
